+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added `operations-for-networks-vsphere-integration.role` file for Aria Operations for Networks to vSphere Integration.
 - Updated `cbo-vsphere-integration.role` to `operations-vsphere-integration.role` so that its more closely aligned to the product.
 - Updated `cbl-vsphere-integration.role` to `operations-for-logs-vsphere-integration.role` so that its more closely aligned to the product.
+- Updated `Export-WsaJsonSpec` to support VMware Cloud Foundation v4.5.1 and Workspace ONE Access
 
 ## v2.2.0 (2023-04-25)